requirements.txt
django~=4.1.7
psycopg2

djangorestframework~=3.14.0
django-filter~=23.2
djangorestframework-simplejwt